Resolve warning, no change.
git-svn-id: https://svn.code.sf.net/p/plisgo/code@184 f301a074-b28b-4858-84c8-f8b36c4887ce

+++ Plisgo/PlisgoFSMenu.cpp
@@ -53,7 +53,7 @@ PlisgoFSMenu::PlisgoFSMenu(IPtrFSIconRegistry FSIcons, const std::wstring& rsPat
 
 				DWORD nWritten = 0;
 
-				WriteFile(m_hSelectionFile, sSelection.c_str(), sSelection.length(), &nWritten, NULL);
+				WriteFile(m_hSelectionFile, sSelection.c_str(), (DWORD)sSelection.length(), &nWritten, NULL);
 
 				WriteFile(m_hSelectionFile, "\r\n", 2, &nWritten, NULL);
 			}
